feat(ui5-writer): abstract functions
