+++ packages/project-access/src/file/index.ts
@@ -1,2 +1,2 @@
 export { fileExists, readFile, readJSON } from './file-access';
-export { findBy, findFiles, findFilesByExtension, findFileUp } from './file-search';
+export { findBy, findFiles, findFilesByExtension, findFileUp, getFilePaths } from './file-search';
