+++ packages/ui5-application-writer/src/data/defaults.ts
@@ -1,3 +1,4 @@
+import { UI5_DEFAULT, getEsmTypesVersion, getTypesVersion } from '@sap-ux/ui5-config';
 import type { App, AppOptions, Package, UI5, UI5Framework } from '../types';
 import versionToManifestDescMapping from '@ui5/manifest/mapping.json'; // from https://github.com/SAP/ui5-manifest/blob/master/mapping.json
 import { getUI5Libs } from './ui5Libs';
@@ -27,42 +28,6 @@ export function packageDefaults(version?: string, description?: string): Partial
             build: 'ui5 build --config=ui5.yaml --clean-dest --dest dist'
         }
     };
-}
-
-/**
- * Merges two objects. All properties from base and from extension will be present.
- * Overlapping properties will be used from extension. Arrays will be concatenated and de-duped.
- *
- * @param base - any object definition
- * @param extension - another object definition
- * @returns - a merged package defintion
- */
-export function mergeObjects<B, E>(base: B, extension: E): B & E {
-    return merge({}, base, extension, (objValue: unknown, srcValue: unknown) => {
-        // merge and de-dup arrays
-        if (objValue instanceof Array && srcValue instanceof Array) {
-            return [...new Set([...objValue, ...srcValue])];
-        } else {
-            return undefined;
-        }
-    });
-}
-
-export const enum UI5_DEFAULT {
-    DEFAULT_UI5_VERSION = '',
-    DEFAULT_LOCAL_UI5_VERSION = '1.95.0',
-    MIN_UI5_VERSION = '1.60.0',
-    MIN_LOCAL_SAPUI5_VERSION = '1.76.0',
-    MIN_LOCAL_OPENUI5_VERSION = '1.52.5',
-    SAPUI5_CDN = 'https://ui5.sap.com',
-    OPENUI5_CDN = 'https://openui5.hana.ondemand.com',
-    TYPES_VERSION_SINCE = '1.76.0',
-    TYPES_VERSION_BEST_MIN = '1.102.0',
-    TYPES_VERSION_PREVIOUS = '1.71.15',
-    TYPES_VERSION_BEST = '1.108.0',
-    ESM_TYPES_VERSION_SINCE = '1.94.0',
-    MANIFEST_VERSION = '1.12.0',
-    BASE_COMPONENT = 'sap/ui/core/UIComponent'
 }
 
 /**
@@ -118,45 +83,6 @@ export function mergeUi5(ui5: Partial<UI5>, options?: Partial<AppOptions>): UI5 
     merged.ui5Libs = getUI5Libs(ui5.ui5Libs);
 
     return Object.assign({}, ui5, merged) as UI5;
-}
-
-/**
- * Get the best types version for the given minUI5Version for https://www.npmjs.com/package/@sapui5/ts-types where specific versions are missing.
- *
- * @param minUI5Version the minimum UI5 version that needs to be supported
- * @returns semantic version representing the types version.
- */
-export function getTypesVersion(minUI5Version?: string) {
-    const version = semVer.coerce(minUI5Version);
-    if (!version) {
-        return `~${UI5_DEFAULT.TYPES_VERSION_BEST}`;
-    } else if (semVer.gte(version, UI5_DEFAULT.TYPES_VERSION_BEST)) {
-        return `~${UI5_DEFAULT.TYPES_VERSION_BEST}`;
-    } else {
-        return semVer.gte(version, UI5_DEFAULT.TYPES_VERSION_SINCE)
-            ? `~${semVer.major(version)}.${semVer.minor(version)}.${semVer.patch(version)}`
-            : UI5_DEFAULT.TYPES_VERSION_PREVIOUS;
-    }
-}
-
-/**
- * Get the best types version for the given minUI5Version within a selective range, starting at 1.90.0
- * for https://www.npmjs.com/package/@sapui5/ts-types-esm
- * For the latest versions the LTS S/4 on-premise version (1.102.x) is used, for anything before we
- * match the versions as far back as available.
- *
- * @param minUI5Version the minimum UI5 version that needs to be supported
- * @returns semantic version representing the types version.
- */
-export function getEsmTypesVersion(minUI5Version?: string) {
-    const version = semVer.coerce(minUI5Version);
-    if (!version || semVer.gte(version, UI5_DEFAULT.TYPES_VERSION_BEST_MIN)) {
-        return `~${UI5_DEFAULT.TYPES_VERSION_BEST}`;
-    } else {
-        return semVer.gte(version, UI5_DEFAULT.ESM_TYPES_VERSION_SINCE)
-            ? `~${semVer.major(version)}.${semVer.minor(version)}.0`
-            : `~${UI5_DEFAULT.ESM_TYPES_VERSION_SINCE}`;
-    }
 }
 
 /**
