+++ packages/ui5-application-writer/src/index.ts
@@ -3,13 +3,12 @@ import { create as createStorage } from 'mem-fs';
 import type { Editor } from 'mem-fs-editor';
 import { create } from 'mem-fs-editor';
 import type { App, AppOptions, Package, UI5 } from './types';
-import { UI5Config } from '@sap-ux/ui5-config';
+import { UI5Config, getEsmTypesVersion } from '@sap-ux/ui5-config';
 import type { Manifest } from '@sap-ux/project-access';
 import { mergeWithDefaults } from './data';
 import { ui5TSSupport } from './data/ui5Libs';
 import { applyOptionalFeatures, enableTypescript as enableTypescriptOption } from './options';
 import { Ui5App } from './types';
-import { getEsmTypesVersion } from './data/defaults';
 
 /**
  * Writes the template to the memfs editor instance.
