+++ packages/ui5-application-writer/test/data.test.ts
@@ -1,6 +1,7 @@
-import { UI5_DEFAULT, mergeUi5, defaultUI5Libs, mergeApp, getSpecTagVersion, mergeObjects } from '../src/data/defaults';
+import { UI5_DEFAULT } from '@sap-ux/ui5-config';
+import { mergeUi5, defaultUI5Libs, mergeApp, getSpecTagVersion } from '../src/data/defaults';
 import { mergeWithDefaults } from '../src/data/index';
-import type { App, Package, UI5, Ui5App } from '../src/types';
+import type { App, UI5, Ui5App } from '../src/types';
 
 const mockSpecVersions = JSON.stringify({ latest: '1.102.3', 'UI5-1.71': '1.71.64', 'UI5-1.92': '1.92.1' });
 jest.mock('child_process', () => ({
@@ -14,50 +15,6 @@ jest.mock('child_process', () => ({
         on: (_event: string, fn: Function) => fn(0)
     })
 }));
-
-describe('mergeObjects', () => {
-    const base: Partial<Package> = {
-        scripts: {
-            first: 'first'
-        },
-        ui5: {
-            dependencies: ['module-1']
-        }
-    };
-
-    test('additional ui5 dependencies (array merge)', () => {
-        const extension: Package = {
-            name: 'test',
-            ui5: {
-                dependencies: ['module-2']
-            }
-        };
-        const merged = mergeObjects(base, extension);
-        expect(merged.ui5?.dependencies).toStrictEqual(['module-1', 'module-2']);
-    });
-
-    test('duplicated ui5 dependencies (array merge)', () => {
-        const extension: Package = {
-            name: 'test',
-            ui5: {
-                dependencies: ['module-1', 'module-2']
-            }
-        };
-        const merged = mergeObjects(base, extension);
-        expect(merged.ui5?.dependencies).toStrictEqual(['module-1', 'module-2']);
-    });
-
-    test('overwrite property', () => {
-        const extension: Package = {
-            name: 'test',
-            scripts: {
-                first: 'second'
-            }
-        };
-        const merged = mergeObjects(base, extension);
-        expect(merged.scripts?.first).toBe(extension.scripts?.first);
-    });
-});
 
 describe('Setting defaults', () => {
     const testData: { input: Partial<UI5>; expected: UI5 }[] = [
