+++ packages/ui5-config/package.json
@@ -31,11 +31,15 @@
         "!dist/**/*.map"
     ],
     "dependencies": {
-        "@sap-ux/yaml": "workspace:*"
+        "@sap-ux/yaml": "workspace:*",
+        "lodash": "4.17.21",
+        "semver": "7.3.5"
     },
     "devDependencies": {
         "mem-fs": "2.1.0",
-        "mem-fs-editor": "9.4.0"
+        "mem-fs-editor": "9.4.0",
+        "@types/lodash": "4.14.176",
+        "@types/semver": "7.3.9"
     },
     "engines": {
         "pnpm": ">=6.26.1 < 7.0.0 || >=7.1.0",
