+++ packages/ui5-config/src/defaults.ts
@@ -1,0 +1,16 @@
+export const enum UI5_DEFAULT {
+    DEFAULT_UI5_VERSION = '',
+    DEFAULT_LOCAL_UI5_VERSION = '1.95.0',
+    MIN_UI5_VERSION = '1.60.0',
+    MIN_LOCAL_SAPUI5_VERSION = '1.76.0',
+    MIN_LOCAL_OPENUI5_VERSION = '1.52.5',
+    SAPUI5_CDN = 'https://ui5.sap.com',
+    OPENUI5_CDN = 'https://openui5.hana.ondemand.com',
+    TYPES_VERSION_SINCE = '1.76.0',
+    TYPES_VERSION_BEST_MIN = '1.102.0',
+    TYPES_VERSION_PREVIOUS = '1.71.15',
+    TYPES_VERSION_BEST = '1.108.0',
+    ESM_TYPES_VERSION_SINCE = '1.94.0',
+    MANIFEST_VERSION = '1.12.0',
+    BASE_COMPONENT = 'sap/ui/core/UIComponent'
+}
